% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irv4.R
\name{irv_pivot_probs_four_cands}
\alias{irv_pivot_probs_four_cands}
\title{Compute pivot probabilities and generate P-matrices for four-candidate IRV elections}
\usage{
irv_pivot_probs_four_cands(sims, n = 1000, reporting = 1)
}
\arguments{
\item{sims}{A tibble of simulations, with an `id` column and
a column for each distinct ballot}

\item{n}{The number of voters.}

\item{reporting}{The level of reporting detail (0: none; 1:
type of pivot event; 2: name of pivot event)}
}
\value{
A list of pivot events, each of which is a list with elements
`integral` and `P`
}
\description{
Compute pivot probabilities and generate P-matrices for four-candidate IRV elections
}
\examples{
sims <- simulate_ordinal_results_from_dirichlet(k = 4, n = 10000)
out <- irv_pivot_probs_four_cands(sims)
out \%>\% combine_P_matrices()
}
