% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plurality.R
\name{plurality_pivot_probs}
\alias{plurality_pivot_probs}
\title{Compute pivot probabilities for a plurality election}
\usage{
plurality_pivot_probs(method = "eggers-vivyan", ...)
}
\arguments{
\item{method}{One of
\itemize{
\item \code{"eggers-vivyan"}, which is based on the Dirichlet
distribution but makes an independence assumption for efficiency -- in brief, it assumes that the probability of $a$ and $b$ tying for first at vote share $x$ is equal to the probability of $a$ and $b$ receiving $x$ times the probability of $c$ receiving no more than $x$ times the probability of $d$ receiving no more than $x$ times the probability of $e$ receiving no more than $x$, etc. (In reality, these events are not independent.)
\item \code{"myatt-fisher"}, which computes the exact (but not normalized) pivot probabilities assuming a Dirichlet distribution and only for the three-candidate case.
\item \code{"simplicial-cubature"}, which uses an adaptive algorithm to integrate the distribution of election results (either Dirichlet or multivariate normal) over the relevant simplices in $\mathbb{R}^k$.
\item \code{"simulation"}, which takes a matrix of simulated election results, or arguments to generate such a matrix from a Dirichlet or multivariate normal distribution, and computes pivot probabilities.
\item \code{"grid-based"}, which constructs a grid of points and computes pivot probabilities using either Dirichlet or multivariate normal parameters.
}}

\item{...}{Other arguments including
\itemize{
\item \code{alpha}, a$k$-length vector of parameters for the Dirichlet
 distribution,
 \item \code{mu}, a $k$-length vector of expected vote shares for the Dirichlet distribution or multivariate normal distribution,
 \item \code{sigma}, a $k$-by-$k$ matrix of variances and covariances for the multivariate normal distribution,
 \item \code{sims} to be used if \code{method = "simulation"},
 \item \code{n} the number of simulations, if \code{method = "simulation"} and not passing \code{sims},
 \item \code{increment}, the space between grid-points if using \code{method = "grid-based"}
 }}
}
\description{
Compute pivot probabilities for a plurality election
}
