% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simulated_election_results.R
\name{simulate_ordinal_results_from_dirichlet}
\alias{simulate_ordinal_results_from_dirichlet}
\title{Generate simulated election results from Dirichlt distribution}
\usage{
simulate_ordinal_results_from_dirichlet(k = 4, n = 10000, alpha = NULL)
}
\arguments{
\item{k}{The number of candidates}

\item{n}{The number of simulated elections to draw}

\item{alpha}{Optional vector of Dirichlet parameters of length `factorial(k)`.
If omitted, random parameters will be drawn.}
}
\value{
A tibble with a column `id` and a column for each unique ballot type.
}
\description{
Generate simulated election results from Dirichlt distribution
}
\examples{
simulate_ordinal_results_from_dirichlet(n = 10)
}
