% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation_based_plurality_pivot_probs}
\alias{simulation_based_plurality_pivot_probs}
\title{Compute pivot probabilities from a matrix of simulated election results}
\usage{
simulation_based_plurality_pivot_probs(
  sims = NULL,
  alpha = NULL,
  mu = NULL,
  precision = NULL,
  sigma = NULL,
  n = 1e+05,
  tol = 0.01,
  cand_names = NULL,
  sep = ""
)
}
\description{
Compute pivot probabilities from a matrix of simulated election results
}
