% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_list_functions.R
\name{election_list_functions}
\alias{election_list_functions}
\alias{plurality_election}
\alias{positional_election}
\alias{irv_election}
\alias{kemeny_young_election}
\title{Create election list for analysis}
\usage{
plurality_election(n = 1000, k = 4, max_pivot_event_degree = 2)

positional_election(n = 1000, s = 0.5)

irv_election(n = 1000, s = 0)

kemeny_young_election(n = 1000)
}
\arguments{
\item{n}{The size of the electorate. This (slightly) affects the
conditions for election events in the election list
(e.g. the region in which one
candidate is more than one vote ahead of another).
\code{n} is also used when computing pivot event probabilities
from an election list in \code{election_event_probabilities()},
where higher \code{n} means lower probability.}

\item{k}{The number of candidates in plurality.}

\item{s}{The value of a second-ranking in a three-candidate
positional election, where a top ranking is worth 1 and a bottom ranking is worth 0.}
}
\description{
An election list represents an election
in a minimal way: electorate size, ballot format, and
rules for determining the winner, electorate size.
It is passed to the \code{election_event_probabilities()}
function to compute the probability of each election event.
}
\details{
An election list contains the electorate size \code{n},
an \code{ordinal}
flag indicating whether this is an ordinal or single-ballot
voting
system, and a list of election \code{events}. For each event,
the election list indicates \itemize{\item conditions under which the event takes place, and \item which candidate wins given that this
event takes place as a function of which additional ballot is submitted.}

These functions produce plain lists, so functions can be written
for further election methods.
}
